+++ attributes/default.rb
@@ -2,11 +2,12 @@
 # TODO(chrislaco) This file needs the stock chef header/copyright
 
 #--------------------
-# keystone auth params
+# keystone auth params and endpoint info
 #--------------------
 default["openstack"]["object-storage"]["service_tenant_name"] = "service"
 default["openstack"]["object-storage"]["service_user"] = "swift"
 default["openstack"]["object-storage"]["service_role"] = "admin"
+default['openstack']['endpoints']['object-storage-api-internal']['uri'] = node['openstack']['endpoints']['object-storage-api']['uri']
 
 #--------------------
 # node/ring settings
@@ -21,7 +22,7 @@ default['openstack']['object-storage']['git_builder_ip'] = '127.0.0.1'
 
 # the release only has any effect on ubuntu, and must be
 # a valid release on http://ubuntu-cloud.archive.canonical.com/ubuntu
-default['openstack']['object-storage']['release'] = 'havana'
+default['openstack']['object-storage']['release'] = 'icehouse'
 
 # we support an optional secret databag where we will retrieve the
 # following attributes overriding any default attributes here